requirements.txt
aiohttp==3.7.2
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
awsebcli==3.19.2
beautifulsoup4==4.9.3
botocore==1.19.25
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.43.0
django-bootstrap4==2.3.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
filelock==3.0.12
future==0.16.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
multidict==5.0.0
oauthlib==3.1.0
pathspec==0.5.9
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1

python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0

PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0

soupsieve==2.0.1
sqlparse==0.4.1
stripe==2.55.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.25.11

wcwidth==0.1.9
whitenoise==5.2.0
yarl==1.6.2
